feat: 复制流水线没有将通知、基础设置一起复制 #2648
