+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/service/PipelineInfoService.kt
@@ -2,10 +2,8 @@ package com.tencent.devops.process.engine.service
 
 import com.tencent.devops.common.api.exception.OperationException
 import com.tencent.devops.common.api.exception.PermissionForbiddenException
-import com.tencent.devops.common.api.util.DateTimeUtil
 import com.tencent.devops.common.api.util.JsonUtil
 import com.tencent.devops.common.auth.api.AuthPermission
-import com.tencent.devops.common.auth.api.pojo.BkAuthGroup
 import com.tencent.devops.common.pipeline.enums.ChannelCode
 import com.tencent.devops.common.pipeline.extend.ModelCheckPlugin
 import com.tencent.devops.common.service.utils.MessageCodeUtil
@@ -13,11 +11,8 @@ import com.tencent.devops.process.constant.ProcessMessageCode.ILLEGAL_PIPELINE_M
 import com.tencent.devops.process.constant.ProcessMessageCode.USER_NEED_PIPELINE_X_PERMISSION
 import com.tencent.devops.process.engine.cfg.PipelineIdGenerator
 import com.tencent.devops.process.permission.PipelinePermissionService
-import com.tencent.devops.process.pojo.pipeline.PipelineSubscriptionType
 import com.tencent.devops.process.pojo.setting.PipelineModelAndSetting
-import com.tencent.devops.process.pojo.setting.PipelineRunLockType
 import com.tencent.devops.process.pojo.setting.PipelineSetting
-import com.tencent.devops.process.pojo.setting.Subscription
 import com.tencent.devops.process.service.PipelineSettingService
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
@@ -45,7 +40,7 @@ class PipelineInfoService @Autowired constructor(
                 permission = AuthPermission.EDIT,
                 message = "用户($userId)无权限在工程($projectId)下导出流水线"
         )
-        val settingInfo = getSettingInfo(projectId, pipelineId, userId)
+        val settingInfo = pipelineSettingService.getSettingInfo(projectId, pipelineId, userId)
         val model = pipelineRepositoryService.getModel(pipelineId)
         if (settingInfo == null || model == null) {
             throw OperationException(MessageCodeUtil.getCodeLanMessage(ILLEGAL_PIPELINE_MODEL_JSON))
@@ -81,7 +76,7 @@ class PipelineInfoService @Autowired constructor(
         val newSetting = PipelineSetting(
                 projectId = projectId,
                 pipelineId = newPipelineId,
-                pipelineName = oldSetting.pipelineName,
+                pipelineName = model.name,
                 desc = oldSetting.desc,
                 successSubscription = oldSetting.successSubscription,
                 failSubscription = oldSetting.failSubscription,
@@ -96,49 +91,6 @@ class PipelineInfoService @Autowired constructor(
         pipelineSettingService.saveSetting(userId, newSetting, false)
 
         return newPipelineId
-    }
-
-    private fun getSettingInfo(projectId: String, pipelineId: String, userId: String): PipelineSetting? {
-        val settingInfo = pipelineSettingService.getSetting(pipelineId) ?: return null
-
-        val hasPermission = pipelinePermissionService.isProjectUser(userId = userId, projectId = projectId, group = BkAuthGroup.MANAGER)
-
-        val successType = settingInfo.successType.split(",").filter { i -> i.isNotBlank() }
-                .map { type -> PipelineSubscriptionType.valueOf(type) }.toSet()
-        val failType = settingInfo.failType.split(",").filter { i -> i.isNotBlank() }
-                .map { type -> PipelineSubscriptionType.valueOf(type) }.toSet()
-
-        return PipelineSetting(
-                projectId = settingInfo.projectId,
-                pipelineId = settingInfo.pipelineId,
-                pipelineName = settingInfo.name,
-                desc = settingInfo.desc,
-                runLockType = PipelineRunLockType.valueOf(settingInfo.runLockType),
-                successSubscription = Subscription(
-                        types = successType,
-                        groups = settingInfo.successGroup.split(",").toSet(),
-                        users = settingInfo.successReceiver,
-                        wechatGroupFlag = settingInfo.successWechatGroupFlag,
-                        wechatGroup = settingInfo.successWechatGroup,
-                        wechatGroupMarkdownFlag = settingInfo.successWechatGroupMarkdownFlag,
-                        detailFlag = settingInfo.successDetailFlag,
-                        content = settingInfo.successContent ?: ""
-                ),
-                failSubscription = Subscription(
-                        types = failType,
-                        groups = settingInfo.failGroup.split(",").toSet(),
-                        users = settingInfo.failReceiver,
-                        wechatGroupFlag = settingInfo.failWechatGroupFlag,
-                        wechatGroup = settingInfo.failWechatGroup,
-                        wechatGroupMarkdownFlag = settingInfo.failWechatGroupMarkdownFlag,
-                        detailFlag = settingInfo.failDetailFlag,
-                        content = settingInfo.failContent ?: ""
-                ),
-                labels = emptyList(),
-                waitQueueTimeMinute = DateTimeUtil.secondToMinute(settingInfo.waitQueueTimeSecond),
-                maxQueueSize = settingInfo.maxQueueSize,
-                hasPermission = hasPermission
-        )
     }
 
     private fun exportModelToFile(modelAndSetting: PipelineModelAndSetting, pipelineName: String): Response {
