+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/service/PipelineService.kt
@@ -477,7 +477,23 @@ class PipelineService @Autowired constructor(
         try {
             val copyMode = Model(name, desc ?: model.desc, model.stages)
             modelCheckPlugin.clearUpModel(copyMode)
-            return createPipeline(userId, projectId, copyMode, channelCode)
+            val newPipelineId = createPipeline(userId, projectId, copyMode, channelCode)
+            val settingInfo = pipelineSettingService.getSettingInfo(projectId, pipelineId, userId)
+            if (settingInfo != null) {
+                // setting pipeline需替换成新流水线的
+                pipelineSettingService.rebuildSetting(
+                        oldSetting = settingInfo!!,
+                        projectId = projectId,
+                        newPipelineId = pipelineId,
+                        pipelineName = name
+                )
+                // 复制setting到新流水线
+                pipelineSettingService.saveSetting(
+                        userId = userId,
+                        setting = settingInfo
+                )
+            }
+            return newPipelineId
         } catch (e: JsonParseException) {
             logger.error("Parse process($pipelineId) fail", e)
             throw ErrorCodeException(
